Finished WallCoinEvent
Punisher Coin: the opponent's obstacles become dangerous! if it collides with the obstacles it loses one life!

Supporter coin: if the player eats this coin, all the walls will disappear.

There is a 10 second timer for both events.

Next option maybe change the sprites for the coins and obstacles? and then go to balancing stuff.